Hexlet

+++ src/main/java/hexlet/code/controller/UserController.java
@@ -4,6 +4,12 @@ import hexlet.code.dto.UserDto;
 import hexlet.code.model.User;
 import hexlet.code.repository.UserRepository;
 import hexlet.code.service.UserService;
+import io.swagger.v3.oas.annotations.Operation;
+import io.swagger.v3.oas.annotations.media.Content;
+import io.swagger.v3.oas.annotations.media.Schema;
+import io.swagger.v3.oas.annotations.responses.ApiResponse;
+import io.swagger.v3.oas.annotations.responses.ApiResponses;
+import org.apache.commons.lang3.NotImplementedException;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.security.access.prepost.PreAuthorize;
 import org.springframework.web.bind.annotation.*;
@@ -22,7 +28,10 @@ public class UserController {
     @Autowired
     private UserRepository userRepository;
 
-
+    @Operation(summary = "Get all users")
+    @ApiResponses(@ApiResponse(responseCode = "200", content =
+    @Content(schema = @Schema(implementation = User.class))
+    ))
     @GetMapping(path = "")
     public Iterable<User> getUsers() {
         return userRepository.findAll();
@@ -35,17 +44,23 @@ public class UserController {
 
     @PutMapping(path = "/{id}")
     @PreAuthorize("""
-            @Repository.findById(#id).get().getEmail() == authentication.getName()
-        """)
+                @userRepository.findById(#id).get().getEmail() == authentication.getName()
+            """)
     public User updateUser(@PathVariable long id, @RequestBody @Valid UserDto userDto) {
         return userService.updateUser(id, userDto);
     }
 
     @DeleteMapping(path = "/{id}")
     @PreAuthorize("""
-            @Repository.findById(#id).get().getEmail() == authentication.getName()
-        """)
-    public void deleteUser(@PathVariable long id) {
+                @userRepository.findById(#id).get().getEmail() == authentication.getName()
+            """)
+    public void deleteUser(@PathVariable final long id) {
+
+        if (!userRepository.findById(id).get().getAuthorTasks().isEmpty() ||
+                !userRepository.findById(id).get().getExecutorTasks().isEmpty()) {
+            throw new NotImplementedException("Пользователь имеет одну из задач, его нельзя удалить");
+        }
+
         userRepository.deleteById(id);
     }
 }
